+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added: `media-feature-value-dollar-variable` rule.
 - Added: `at-import-partial-extension-blacklist` rule.
 - Added: `at-import-partial-extension-whitelist` rule.
+- Deprecated: `at-import-no-partial-extension` rule.
 
 # 1.1.1